+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `debounce()` function in `functions.js`
 - `$.debounce()`
 - `loadImage()` function to load images after they have been fetched
+- Polyfill for `img.complete` and `img.decode()`
 
 ### Changed
 - Mark several functions and methods as deprecated, for removal in v3.0.0